requirements.txt
# to install run
# sudo pip install --upgrade -r requirements.txt

#django>=1.3
http://www.djangoproject.com/download/1.3-rc-1/tarball/

fabric
oauth2
django-uni-form
django-extensions
django-openid-auth
django-openid-consumer
django-debug-toolbar
django-email-confirmation
-e git+http://github.com/facebook/python-sdk.git#egg=facebook-sdk
django-allauth
PIL # Python Imaging Library
djangorestframework
cherrypy
django-cpserver
